requirements.txt
# General dependencies
python-dotenv
requests
tenacity
numpy
pandas
sqlalchemy
tqdm

# Install with conda:
# conda install psycopg2

# System dependencies (instructions only, not processed by pip)
# For Linux:
#   sudo apt install -y python3-dev libpq-dev unixodbc-dev
# For macOS:
#   brew install postgresql unixodbc
# For Windows:
#   Download and install ODBC Driver for SQL Server:
#   https://learn.microsoft.com/en-us/sql/connect/odbc/download-odbc-driver-for-sql-server
